🐛 (ControllerExceptionHandler.java): Adiciona tratamento de exceção para DatabaseException 🐛 (ProductService.java): Adiciona tratamento de exceção para EntityNotFoundException e DataIntegrityViolationException ✨ (DatabaseException.java): Adiciona nova exceção para erros de banco de dados
O tratamento de exceções foi adicionado para melhorar a robustez do sistema. Agora, o sistema pode lidar adequadamente com exceções específicas, como DatabaseException, EntityNotFoundException e DataIntegrityViolationException, fornecendo mensagens de erro mais claras e precisas para os usuários finais. Isso ajuda a melhorar a experiência do usuário e a facilitar a depuração de problemas relacionados ao banco de dados.

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,10 +3,14 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -45,16 +49,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
 
-        Product updated = productRepository.save(entity);
-        return new ProductDTO(updated);
+            Product updated = productRepository.save(entity);
+            return new ProductDTO(updated);
+        } catch (EntityNotFoundException ex) {
+            throw new ResourceNotFoundException("Product not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id) {
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException ex) {
+            throw new DatabaseException("Referential integrity failure");
+        }
     }
 
     private static void copyDtoToEntity(ProductDTO dto, Product entity) {
